Refactoring TAMPI and modifying the activation of its modes.
Internal arrays of requests are not increased dynamically and
all alive requests are kept consecutive. Ticket allocator is no
longer needed. The task-based runtime system API is also modified.

Now the blocking mode of TAMPI is enabled when requesting the
MPI_TASK_MULTIPLE threading level. The non-blocking mode is
automatically enabled when MPI_THREAD_MULTIPLE or a greater
threading level is provided.

Fix #12 and #15.