Replacing TAMPI_POLLING_FREQUENCY by TAMPI_POLLING_PERIOD
The TAMPI_POLLING_FREQUENCY envar is replaced by TAMPI_POLLING_PERIOD.
The first one is deprecated but it is still considered. Setting both
envars results in undefined behavior.

Notice that this envar is ignored when TAMPI relies on the polling
services API. This is the case of the LLVM OpenMP runtime system.