Implementing an initial dynamic polling period
The TAMPI_POLLING_PERIOD envar is extended to support additional information.
The envar follows the format '<min_period>[:<max_period>]`. If the user only
defines the minimum period, the polling period is static, as originally. If
both minimum and maximum periods are defined, the polling dynamically changes
depending on the completed and pending requests.

The dynamic polling is not documented yet since it is an experimental feature.