Dropping support for OmpSs-2 2020.06
This commit drops the support for OmpSs-2 2020.06. That version has a
bug in nanos6_spawn_function since it does not spawn an independent task
in all cases. As a workaround, we modified TAMPI to spawn the task from
an external thread and ensure the spawned task is independent. However,
this external thread produces issues with the ovni tracing. This commit
just removes the creation of that thread.