Renaming, improving and extending boost lockfree queues
LockFreeQueue is renamed to BoostLockFreeQueue. Pushing to the
queue no longer requires a progress function.

By default, tasks try to push to queues until there is available space
in the queue. The TAMPI_QUEUES_FULL_FAILURE envar controls whether the
execution should about when queues become full. The envar is disabled
by default.

Enabling this envar is useful in performance runs to keep control of when
queues become full.