Restoring TAMPI_POLLING_FREQUENCY as deprecated for compatibility
The envar TAMPI_POLLING_FREQUENCY keeps deprecated and will be
removed soon in future versions. Warn the user if the envar is
defined. The currently valid envar is TAMPI_POLLING_PERIOD The
value of TAMPI_POLLING_FREQUENCY is only considered when the
TAMPI_POLLING_PERIOD is not defined

It also removes an experimental but obsolete feature