Enhancing and protecting MPI and TAMPI initialization
This commit introduces a mutex to protect the initalization and
finalization of the library. The initialization is divided into
two parts: preinitialization and initialization. The first sets
up the MPI-related environment and the seconds actually initializes
the TAMPI library, such as the runtime symbols, polling task and
the instrumentation

Consulting whether the blocking or non-blocking modes are enabled
does not involve acquiring the lock. Atomic operations are used
in that case