Fixing CPU binding in SLURM-based tests
Beginning with 22.05, srun does not inherit the --cpus-per-task value requested by
the salloc or sbatch command. This can provoke the MPI processes to use a single
cpu, damaging performance and testing purposes.

This commit fixes the testing script to use the requested configuration from the
current sbatch/salloc session. The script reads the number of nodes, the number of
processes, and the number of cpus per process from the SLURM environment. Note that
the SLURM session must be allocated the options --nodes (-N), --ntasks (-n), and
--cpus-per-task (-c). The envars read by the script are the following:

  - SLURM_JOB_NUM_NODES
  - SLURM_NTASKS
  - SLURM_CPUS_PER_TASK