Require TAMPI model with ovni_thread_require()
The ovni emulator needs to know beforehand the list of models to be
enabled (and their versions) for the emulation. As the thread streams
are managed by the runtime, we don't have a way of running a hook per
thread, so we use a thread local flag and call the require function on
the first event to be emitted in each thread.