Implementing scalable allocator for Operations
The allocator benefits from the condition that user tasks allocate
Operations, and these are only processed by the polling task. The
allocator has a Boost SPSC queue with a lock protecting the consumer
side only. This queue holds all the allocated objects initially.

Each CPU has a small cache (stack) of objects, which is only accessed
by the same CPU, so they do not require inter-thread synchronization.
The caches are empty initially. When a user task needs an Operation
object, it first checks the cache for a fast allocation. Otherwise,
when the cache is empty, the CPU retrives N Operations from the SPSC
queue and stores them in the cache.

The polling task processes the Operations, and after that, it returns
the objects to the SPSC queue (without acquiring any lock).

This design should be compatible with other types of objects from any
other task-aware library.